Add LTS streams and update all
Ran a full update with two streams enabled (4.4, 4.9). This is first push of
any scans of LTS streams. In future we plan to do as many LTS streams as
wanted but these seem like the best two to start testing on. See the readme
for how to understand these. Since they are autogenerated there still needs
to be some more validation/testing to insure accuracy but i rather release
them earlier rather than later.