Add vendor specific handling and update all
This adds information in the kernel_cves.json to indicate whether the CVE is
vendor specific or not. This is used by the autogenerator to avoid printing
CVEs that do not apply to the upstream kernel. Not all CVEs have been checked
for vendor specific information yet. This will be a slow and mostly manual
process.

This closes #5