Update 08Mar17 and backport solution
This includes the daily update (the first) for 08Mar17.

This also is the first run through of the backporting checking
solution. We now flag backported issues and check for the introducing
commit, if the introducing commit is present we then check for the
fixing commit as always. Prior to this commit some fixing commits
were set at the version they were considered fixed, with different
commit messages of the actual fix for the commit. These will stay,
but going forward (now that they are tracked better) we will use
matching cmt hashes and cmt msges. This closes #6.